rename sm90 megamoe kernel: drop cooperative suffix, converge API, clean comments
- sm90_fp8_mega_moe_cooperative.{cuh,hpp} -> sm90_fp8_mega_moe.{cuh,hpp};
  kernel/runtime/launcher/heuristic symbols renamed accordingly
- remove duplicated fp8_mega_moe_cooperative C++/Python entry and the
  DG_SM90_MOE_KERNEL routing in tests/bench
- compress version-history comment to one-line JIT cache key (v30); comments
  trimmed to SM100 style; fix stale per-64 / M-split wording
- drop unused kSwizzleCDMode; cd_stages default 1 -> 2; order_barriers kept
  (reserved for L2 pingpong)
