+++ csrc/apis/mega.hpp
@@ -9,7 +9,7 @@
 #endif
 #include "../jit/device_runtime.hpp"
 #include "../jit_kernels/impls/sm100_fp8_fp4_mega_moe.hpp"
-#include "../jit_kernels/impls/sm90_fp8_mega_moe_cooperative.hpp"
+#include "../jit_kernels/impls/sm90_fp8_mega_moe.hpp"
 
 namespace deep_gemm::mega {
 
@@ -342,123 +342,8 @@ static void fp8_mega_moe(
     // Already registered tensors
     const auto [x, x_sf, topk_idx, topk_weights, l1_acts, l1_acts_sf, l2_acts, l2_acts_sf] = slice(sym_buffer);
 
-    // Single cooperative kernel for all token counts: the N-split cooperative
-    // kernel (BLOCK_M=64, BLOCK_N=256, two math warpgroups split the 256-wide N
-    // tile into two 128-wide halves and share one A-tile load) is the only SM90
-    // mega-MoE path. It quantizes the L2-input activations at per-128 K to match
-    // the standard DeepEP runner. (The former pingpong kernel and the
-    // token-count routing threshold were removed.)
-    sm90_fp8_mega_moe_cooperative(y,
-                     l1_acts, l1_acts_sf,
-                     l2_acts, l2_acts_sf,
-                     l1_weights, l2_weights,
-                     l1_weights_sf, l2_weights_sf,
-                     cumulative_local_expert_recv_stats,
-                     sym_buffer_ptrs,
-                     rank_idx, num_max_tokens_per_rank,
-                     num_experts_per_rank,
-                     num_tokens, num_topk,
-                     hidden, intermediate_hidden,
-                     activation_clamp, fast_math);
-
-    if (get_env<int>("DG_COMM_KERNEL_DEBUG"))
-        sym_buffer.zero_();
-}
-
-// SM90 (Hopper) FP8 MegaMoE entry point — COOPERATIVE implementation.
-// Single-launch fused N-split cooperative kernel (block_m=64, block_n=256): two
-// math warpgroups split the 256-wide N tile into two 128-wide halves, share one
-// A-tile load, and quantize the L2-input activations at per-128 K. Body is
-// identical to `fp8_mega_moe` except this is the explicit (non-auto) entry.
-// Exposed to Python as `fp8_mega_moe_cooperative`.
-static void fp8_mega_moe_cooperative(
-    const torch::Tensor& y,
-    const std::tuple<torch::Tensor, torch::Tensor>& l1_weights_tuple,
-    const std::tuple<torch::Tensor, torch::Tensor>& l2_weights_tuple,
-    const std::optional<torch::Tensor>& cumulative_local_expert_recv_stats,
-    const torch::Tensor& sym_buffer,
-    const std::vector<int64_t>& sym_buffer_ptrs, const int& rank_idx,
-    const int& num_max_tokens_per_rank,
-    const int& num_experts, const int& num_topk,
-    const std::tuple<int, int, int>& recipe,
-    const std::string& activation,
-    const std::optional<float>& activation_clamp_opt,
-    const bool& fast_math
-) {
-    const auto [l1_weights, l1_weights_sf] = l1_weights_tuple;
-    const auto [l2_weights, l2_weights_sf] = l2_weights_tuple;
-
-    // Architecture check
-    const auto arch_major = device_runtime->get_arch_major();
-    DG_HOST_ASSERT(arch_major == 9);
-
-    // Config checks: SM90 uses block (128, 128) float SF for weights,
-    // per-token per-128-K float SF for activations.
-    const auto num_tokens = static_cast<int>(y.size(0));
-    const auto [rm, rn, rk] = recipe;
-    DG_HOST_ASSERT(rm == 128 and rn == 128 and rk == 128);
-    DG_HOST_ASSERT(activation == "swiglu");
-
-    // Activation checks
-    const auto activation_clamp =
-        activation_clamp_opt.value_or(std::numeric_limits<float>::infinity());
-    DG_HOST_ASSERT(activation_clamp >= 0);
-
-    // Tensor checks: SM90 weights must be FP8 e4m3, K-major
-    DG_HOST_ASSERT(get_major_type_ab(l1_weights) == cute::UMMA::Major::K);
-    DG_HOST_ASSERT(get_major_type_ab(l2_weights) == cute::UMMA::Major::K);
-    DG_HOST_ASSERT(l1_weights.scalar_type() == torch::kFloat8_e4m3fn);
-    DG_HOST_ASSERT(l2_weights.scalar_type() == torch::kFloat8_e4m3fn);
-    const auto [num_experts_per_rank, intermediate_hidden_2, hidden] = get_shape<3>(l1_weights);
-    const auto [num_experts_per_rank_, hidden_, intermediate_hidden] = get_shape<3>(l2_weights);
-    DG_HOST_ASSERT(num_tokens <= num_max_tokens_per_rank);
-    DG_HOST_ASSERT(num_experts_per_rank == num_experts_per_rank_);
-    DG_HOST_ASSERT(hidden == hidden_);
-    DG_HOST_ASSERT(intermediate_hidden_2 == 2 * intermediate_hidden);
-    DG_HOST_ASSERT(l1_weights.is_contiguous() and l2_weights.is_contiguous());
-
-    // Shape constraints required by the SM90 kernel (BLOCK_N=256):
-    //   * L2_SHAPE_N = hidden must be a multiple of BLOCK_N=256 (scheduler tiling).
-    //   * L1_SHAPE_N = 2*intermediate_hidden must be a multiple of 256, i.e.
-    //     intermediate_hidden % 128 == 0.
-    //   * `l2_arrival_mask` is uint64, with one bit per L1-output N-block. Each
-    //     block spans 256 weight cols (= 128 post-SwiGLU cols), so
-    //     `kNumL1BlockNs = 2*intermediate_hidden / 256 = intermediate_hidden / 128`
-    //     must be ≤ 64.
-    DG_HOST_ASSERT(hidden % 256 == 0 and intermediate_hidden % 128 == 0);
-    DG_HOST_ASSERT(intermediate_hidden / 128 <= 64);
-
-    // Check weight SF layout (block (128, 128) float, MN-major; not TMA-loaded
-    // so no TMA-stride alignment is required, but we do require contiguity in
-    // the K-direction within each expert).
-    constexpr int kGranMN = 128, kGranK = 128;
-    check_sf_layout(l1_weights_sf, intermediate_hidden * 2, hidden, kGranMN, kGranK,
-                    num_experts_per_rank, false, true, torch::kFloat);
-    check_sf_layout(l2_weights_sf, hidden, intermediate_hidden, kGranMN, kGranK,
-                    num_experts_per_rank, false, true, torch::kFloat);
-
-    // Check stats counter
-    if (cumulative_local_expert_recv_stats.has_value()) {
-        DG_HOST_ASSERT(cumulative_local_expert_recv_stats->scalar_type() == torch::kInt);
-        DG_HOST_ASSERT(cumulative_local_expert_recv_stats->numel() == num_experts_per_rank);
-        DG_HOST_ASSERT(cumulative_local_expert_recv_stats->is_contiguous());
-    }
-
-    // Check buffer bytes
-    const auto num_ranks = static_cast<int>(sym_buffer_ptrs.size());
-    const auto num_experts_ = num_experts_per_rank * num_ranks;
-    const auto [num_required_bytes, slice] = get_symm_buffer_size_for_mega_moe(
-        num_ranks, num_experts,
-        num_max_tokens_per_rank, num_topk,
-        hidden, intermediate_hidden,
-        true, activation);
-    DG_HOST_ASSERT(sym_buffer.nbytes() >= static_cast<size_t>(num_required_bytes));
-    DG_HOST_ASSERT(num_experts == num_experts_);
-
-    // Already registered tensors
-    const auto [x, x_sf, topk_idx, topk_weights, l1_acts, l1_acts_sf, l2_acts, l2_acts_sf] = slice(sym_buffer);
-
-    sm90_fp8_mega_moe_cooperative(y,
+    // Single N-split kernel (BLOCK_M=64, BLOCK_N=256) for all token counts
+    sm90_fp8_mega_moe(y,
                      l1_acts, l1_acts_sf,
                      l2_acts, l2_acts_sf,
                      l1_weights, l2_weights,
@@ -481,7 +366,6 @@ static void register_apis(pybind11::module_& m) {
     m.def("get_symm_buffer_size_for_mega_moe", &get_symm_buffer_size_for_mega_moe);
     m.def("fp8_fp4_mega_moe", &fp8_fp4_mega_moe);
     m.def("fp8_mega_moe", &fp8_mega_moe);
-    m.def("fp8_mega_moe_cooperative", &fp8_mega_moe_cooperative);
 #endif
 }
