+++ deep_gemm/__init__.py
@@ -88,7 +88,6 @@ from .mega import (
     transform_weights_for_mega_moe_sm90,
     fp8_fp4_mega_moe,
     fp8_mega_moe,
-    fp8_mega_moe_cooperative,
 )
 
 # Some utils
